smaller refactor
