+++ orchestrator/src/client/components/llmmodelconfiguration/LlmModelConfiguration.tsx
@@ -8,7 +8,6 @@ import {
   LLM_PROVIDER_LABELS,
   LLM_PROVIDERS,
   type LlmProviderId,
-  normalizeLlmProvider,
   supportsLlmModelSuggestions,
 } from "@client/pages/settings/utils";
 import { getDefaultModelForProvider } from "@shared/settings-registry";
@@ -19,7 +18,6 @@ import type {
 } from "@shared/types";
 import type React from "react";
 import { useDeferredValue, useEffect, useState } from "react";
-import { SearchableDropdown } from "@/components/ui/searchable-dropdown";
 import {
   Select,
   SelectContent,
@@ -28,6 +26,9 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Separator } from "@/components/ui/separator";
+import PurposeOverrideCard from "@client/components/llmmodelconfiguration/PurposeOverrideCard";
+import { buildModelOptions, LLM_PURPOSES, renderKeyHelper } from "./llm-model-configuration-helpers";
+import ModelField from "./ModelField";
 
 type TextFieldBinding = {
   value: string;
@@ -350,7 +351,7 @@ export function LlmModelConfiguration({
           <Separator />
 
           <div className="space-y-4">
-            <div className="text-sm font-medium">Purpose Overrides</div>
+            <div className="text-sm font-medium">Purpose-specific Model Overrides</div>
 
             {purposeOverrides ? (
               <div className="grid gap-4 lg:grid-cols-3">
@@ -457,384 +458,4 @@ export function LlmModelConfiguration({
       ) : null}
     </div>
   );
-}
-
-function ModelField({
-  id,
-  label,
-  value,
-  onChange,
-  error,
-  supportsModelSuggestions,
-  options,
-  placeholder,
-  helper,
-  current,
-  disabled,
-}: {
-  id: string;
-  label: string;
-  value: string;
-  onChange: (value: string) => void;
-  error?: string;
-  supportsModelSuggestions: boolean;
-  options: Array<{ value: string; label: string; searchText: string }>;
-  placeholder: string;
-  helper?: React.ReactNode;
-  current: string;
-  disabled: boolean;
-}) {
-  if (supportsModelSuggestions) {
-    return (
-      <div className="space-y-2">
-        <label htmlFor={id} className="text-sm font-medium">
-          {label}
-        </label>
-        <SearchableDropdown
-          inputId={id}
-          value={value}
-          options={options}
-          onValueChange={onChange}
-          placeholder={placeholder}
-          searchPlaceholder="Search models..."
-          emptyText="No models found."
-          ariaLabel={label}
-          disabled={disabled}
-          triggerClassName="h-9 w-full justify-between rounded-md border border-input bg-transparent px-3 text-sm font-normal shadow-sm"
-          contentClassName="w-[var(--radix-popover-trigger-width)] border-border bg-popover p-0"
-          listClassName="max-h-64"
-        />
-        {error ? <p className="text-xs text-destructive">{error}</p> : null}
-        {helper ? (
-          <div className="text-xs text-muted-foreground">{helper}</div>
-        ) : null}
-        <div className="text-xs text-muted-foreground">
-          Current: <span className="font-mono">{current}</span>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <SettingsInput
-      label={label}
-      inputProps={{
-        name: id,
-        value,
-        onChange: (event) => onChange(event.target.value),
-      }}
-      placeholder={placeholder}
-      disabled={disabled}
-      error={error}
-      helper={helper}
-      current={current}
-    />
-  );
-}
-
-const LLM_PURPOSES: Array<{
-  id: LlmPurpose;
-  label: string;
-  description: string;
-}> = [
-  {
-    id: "scoring",
-    label: "Scoring",
-    description: "Job fit, brief extraction, and ranking decisions.",
-  },
-  {
-    id: "tailoring",
-    label: "Tailoring",
-    description: "Resume summaries, skills, and tailoring generation.",
-  },
-  {
-    id: "projectSelection",
-    label: "Project selection",
-    description: "Choosing resume projects for a job.",
-  },
-];
-
-function PurposeOverrideCard({
-  purpose,
-  label,
-  description,
-  defaultProvider,
-  defaultModel,
-  defaultBaseUrl,
-  defaultApiKeyHint,
-  value,
-  apiKeyValue,
-  apiKeyHint,
-  currentModel,
-  disabled,
-  onChange,
-  onApiKeyChange,
-}: {
-  purpose: LlmPurpose;
-  label: string;
-  description: string;
-  defaultProvider: LlmProviderId;
-  defaultModel: string;
-  defaultBaseUrl: string;
-  defaultApiKeyHint: string | null;
-  value?: LlmPurposeOverrides[LlmPurpose];
-  apiKeyValue: string;
-  apiKeyHint: string | null;
-  currentModel: string;
-  disabled: boolean;
-  onChange: (
-    purpose: LlmPurpose,
-    field: "provider" | "baseUrl" | "model",
-    value: string | null,
-  ) => void;
-  onApiKeyChange: (purpose: LlmPurpose, value: string) => void;
-}) {
-  const selectedProvider = value?.provider
-    ? normalizeLlmProvider(value.provider)
-    : defaultProvider;
-  const hasProviderOverride = Boolean(value?.provider);
-  const providerConfig = getLlmProviderConfig(selectedProvider);
-  const supportsModelSuggestions =
-    supportsLlmModelSuggestions(selectedProvider);
-  const [availableModels, setAvailableModels] = useState<string[]>([]);
-  const [modelsError, setModelsError] = useState<string | null>(null);
-  const [isLoadingModels, setIsLoadingModels] = useState(false);
-  const baseUrlValue = value?.baseUrl ?? "";
-  const modelValue = value?.model ?? "";
-  const effectiveBaseUrl = baseUrlValue || defaultBaseUrl;
-  const hasSavedKey = Boolean(apiKeyHint || defaultApiKeyHint);
-  const keyHint = apiKeyHint
-    ? formatSecretHint(apiKeyHint)
-    : defaultApiKeyHint
-      ? `inherits ${formatSecretHint(defaultApiKeyHint)}`
-      : "Not set";
-  const modelOptions = buildModelOptions({
-    models: availableModels,
-    emptyLabel: "Inherit purpose default",
-    emptyValue: "",
-    fallbackValue: modelValue,
-  });
-
-  useEffect(() => {
-    if (!supportsModelSuggestions) {
-      setAvailableModels([]);
-      setModelsError(null);
-      setIsLoadingModels(false);
-      return;
-    }
-
-    if (providerConfig.showApiKey && !apiKeyValue.trim() && !hasSavedKey) {
-      setAvailableModels([]);
-      setModelsError(null);
-      setIsLoadingModels(false);
-      return;
-    }
-
-    let cancelled = false;
-    setIsLoadingModels(true);
-    setModelsError(null);
-
-    void api
-      .getLlmModels({
-        provider: selectedProvider,
-        baseUrl: providerConfig.showBaseUrl
-          ? baseUrlValue.trim() || undefined
-          : undefined,
-        apiKey: providerConfig.showApiKey
-          ? apiKeyValue.trim() || undefined
-          : undefined,
-        purpose,
-      })
-      .then((models) => {
-        if (cancelled) return;
-        setAvailableModels(models);
-      })
-      .catch((error) => {
-        if (cancelled) return;
-        setAvailableModels([]);
-        setModelsError(
-          error instanceof Error ? error.message : "Failed to load models.",
-        );
-      })
-      .finally(() => {
-        if (cancelled) return;
-        setIsLoadingModels(false);
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [
-    apiKeyValue,
-    baseUrlValue,
-    hasSavedKey,
-    providerConfig.showApiKey,
-    providerConfig.showBaseUrl,
-    purpose,
-    selectedProvider,
-    supportsModelSuggestions,
-  ]);
-
-  const modelHelper = supportsModelSuggestions
-    ? isLoadingModels
-      ? "Loading available models..."
-      : modelsError
-        ? modelsError
-        : "Leave blank to inherit the purpose default."
-    : `Type the exact model name, or leave blank to use ${providerConfig.label}.`;
-
-  return (
-    <div className="space-y-4 rounded-lg border border-border bg-card p-4">
-      <div className="space-y-1">
-        <div className="text-sm font-medium">{label}</div>
-        <div className="text-xs text-muted-foreground">{description}</div>
-      </div>
-
-      <div className="space-y-2">
-        <label htmlFor={`${purpose}-provider`} className="text-sm font-medium">
-          Runtime
-        </label>
-        <Select
-          value={value?.provider ?? "__inherit__"}
-          onValueChange={(nextValue) => {
-            onChange(
-              purpose,
-              "provider",
-              nextValue === "__inherit__" ? null : nextValue,
-            );
-            onChange(purpose, "baseUrl", null);
-            onChange(purpose, "model", null);
-          }}
-          disabled={disabled}
-        >
-          <SelectTrigger id={`${purpose}-provider`} className="h-9">
-            <SelectValue placeholder="Inherit default" />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="__inherit__">Inherit default</SelectItem>
-            {LLM_PROVIDERS.map((provider) => (
-              <SelectItem key={provider} value={provider}>
-                {LLM_PROVIDER_LABELS[provider]}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-        <div className="text-xs text-muted-foreground">
-          Current:{" "}
-          <span className="font-mono">
-            {hasProviderOverride
-              ? selectedProvider
-              : `inherits ${defaultProvider}`}
-          </span>
-        </div>
-      </div>
-
-      {hasProviderOverride && providerConfig.showBaseUrl ? (
-        <SettingsInput
-          label="Base URL"
-          inputProps={{
-            name: `${purpose}.baseUrl`,
-            value: baseUrlValue,
-            onChange: (event) =>
-              onChange(purpose, "baseUrl", event.target.value),
-          }}
-          placeholder={providerConfig.baseUrlPlaceholder}
-          disabled={disabled}
-          helper={providerConfig.baseUrlHelper}
-          current={effectiveBaseUrl}
-        />
-      ) : null}
-
-      {hasProviderOverride && providerConfig.showApiKey ? (
-        <SettingsInput
-          label="API key"
-          inputProps={{
-            name: `${purpose}.apiKey`,
-            value: apiKeyValue,
-            onChange: (event) => onApiKeyChange(purpose, event.target.value),
-          }}
-          type="password"
-          placeholder="Paste a purpose key"
-          disabled={disabled}
-          helper={renderKeyHelper(
-            providerConfig.keyHelperText,
-            providerConfig.keyHelperHref,
-            hasSavedKey,
-          )}
-          current={keyHint}
-        />
-      ) : null}
-
-      <ModelField
-        id={`${purpose}-model`}
-        label="Model"
-        value={modelValue}
-        onChange={(nextValue) => onChange(purpose, "model", nextValue)}
-        supportsModelSuggestions={supportsModelSuggestions}
-        options={modelOptions}
-        placeholder={defaultModel || "Inherit model"}
-        helper={modelHelper}
-        current={currentModel}
-        disabled={disabled || isLoadingModels}
-      />
-    </div>
-  );
-}
-
-function renderKeyHelper(
-  helperText: string,
-  helperHref: string | null,
-  keepSavedKey: boolean,
-) {
-  return (
-    <>
-      {helperHref ? (
-        <a
-          href={helperHref}
-          target="_blank"
-          rel="noopener noreferrer"
-          className="underline decoration-border underline-offset-4 transition-colors hover:text-foreground"
-        >
-          {helperText}
-        </a>
-      ) : (
-        helperText
-      )}
-      {keepSavedKey ? ". Leave blank to keep the saved key." : null}
-    </>
-  );
-}
-
-function buildModelOptions(input: {
-  models: string[];
-  emptyLabel: string;
-  emptyValue: string;
-  fallbackValue?: string;
-}) {
-  const options = [
-    {
-      value: input.emptyValue,
-      label: input.emptyLabel,
-      searchText: input.emptyLabel,
-    },
-    ...input.models.map((model) => ({
-      value: model,
-      label: model,
-      searchText: model,
-    })),
-  ];
-
-  const fallbackValue = input.fallbackValue?.trim();
-  if (
-    fallbackValue &&
-    !options.some((option) => option.value === fallbackValue)
-  ) {
-    options.unshift({
-      value: fallbackValue,
-      label: fallbackValue,
-      searchText: `${fallbackValue} custom`,
-    });
-  }
-
-  return options;
 }
